NXP Debian Linux SDK 25.12 Release for i.MX and Layerscape MPU platforms
- Upgrade to Debian13
- Upgrade NXP-specific components based on lf-6.12.20-2.0.0 for BSP, kernel
- Cache all the source code to speed up the compiling next time
- Each board has its own specific rootfs
- Add support of imx95evk, imx95frdm and imx8qmmek boards

+++ src/apps/multimedia/basler_camera.mk
@@ -12,8 +12,10 @@ ifeq ($(CONFIG_BASLER_CAMERA),y)
 	 $(call fbprint_b,"basler_camera") && \
 	 cd $(MMDIR) && \
 	 if [ ! -d $(MMDIR)/basler_camera ]; then \
-	     wget -q $(repo_basler_camera_bin_url) -O basler_camera.bin $(LOG_MUTE) && \
-	     chmod +x basler_camera.bin && ./basler_camera.bin --auto-accept $(LOG_MUTE) && \
+		 rm -rf basler_camera*; \
+	     $(WGET) $(repo_basler_camera_bin_url) -O basler_camera.bin $(LOG_MUTE); \
+		 [ $$? -ne 0 ] && { echo "Downloading $(repo_basler_camera_bin_url) failed."; exit 1; } || \
+	     chmod +x basler_camera.bin && ./basler_camera.bin --auto-accept --force $(LOG_MUTE); \
 	     mv basler-camera-* basler_camera && rm -f basler_camera.bin; \
 	 fi && \
 	 cd basler_camera && \
